line plot fulling working -- others should be easier from here on out..

+++ plot/plotter.go
@@ -14,5 +14,5 @@ type Plotter interface {
 // DataRanger wraps the DataRange method.
 type DataRanger interface {
 	// DataRange returns the range of X and Y values.
-	DataRange() (xmin, xmax, ymin, ymax float64)
+	DataRange() (xmin, xmax, ymin, ymax float32)
 }
